cloud-storage/requirements.txt
protobuf==4.21.9
wget==3.2
google-cloud-storage==2.6.0